refactor: move SocketMocketCore from mocket.utils to mocket.io
